moved launch editor to separate class
